perf: The original function simply passes through to execute_kernel with static_cast<uint64_t> on n_rows and n_progs. Since the function is mainly a thin wrapper around execute_kernel, the optimization focuses on reducing unnecessary operations. The static_cast<uint64_t> operations are redundant as the execute_kernel function can handle the implicit conversion. Removing these explicit casts simplifies the code path and reduces unnecessary operations.
Optimized program.cpp

+++ src/program.cpp
@@ -105,8 +105,7 @@ void compute_metric(int n_rows, int n_progs, const float *y,
 
 void execute(const program_t &d_progs, const int n_rows, const int n_progs,
              const float *data, float *y_pred) {
-  execute_kernel(d_progs, data, y_pred, static_cast<uint64_t>(n_rows),
-                 static_cast<uint64_t>(n_progs));
+  execute_kernel(d_progs, data, y_pred, n_rows, n_progs);
 }
 
 void find_fitness(program_t d_prog, float *score, const param &params,
